# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Beautique Store Backend API
  description: |
    RESTful API for Beautique Store e-commerce platform.

    ## Authentication
    - Public endpoints: No authentication required
    - Admin endpoints (/api/admin/*): JWT Bearer token required

    ## Payment Workflow
    Phase 1 uses manual payment verification via WhatsApp.
    No payment gateway integration.
  version: 1.0.0
  contact:
    name: Beautique Store
    email: admin@beautique.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://beautique-api.hf.space
    description: Production server (Hugging Face Spaces)
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List all products
      description: Returns paginated list of active products with optional filters
      operationId: listProducts
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ProductCategory'
          description: Filter by category
        - name: search
          in: query
          schema:
            type: string
          description: Search in name and description
        - name: minPrice
          in: query
          schema:
            type: number
            minimum: 0
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
            maximum: 1000000
          description: Maximum price filter
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 50
          description: Items per page
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product details
      description: Returns complete product information by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/products:
    post:
      tags: [Admin Products]
      summary: Create product
      description: Creates a new product
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/admin/products/{id}:
    put:
      tags: [Admin Products]
      summary: Update product
      description: Updates an existing product
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Admin Products]
      summary: Delete product
      description: Soft-deletes a product (sets is_active=false)
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/products/{id}/images:
    post:
      tags: [Admin Products]
      summary: Upload product images
      description: Uploads images to Cloudinary and adds URLs to product
      operationId: uploadProductImages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
      responses:
        '200':
          description: Images uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  images:
                    type: array
                    items:
                      type: string
                      format: uri
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCategory:
      type: string
      enum:
        - Maxi
        - Lehanga Choli
        - Long Shirt
        - Shalwar Kameez
        - Gharara
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 3
          maxLength: 255
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          type: number
          minimum: 0
          maximum: 1000000
        description:
          type: string
          minLength: 10
          maxLength: 2000
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        sizes:
          type: array
          items:
            type: string
            enum: [S, M, L, XL, XXL]
          minItems: 1
        colors:
          type: array
          items:
            type: string
          minItems: 1
        is_bestseller:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductCreate:
      type: object
      required:
        - name
        - category
        - price
        - description
        - images
        - sizes
        - colors
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          type: number
          minimum: 0
          maximum: 1000000
        description:
          type: string
          minLength: 10
          maxLength: 2000
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        sizes:
          type: array
          items:
            type: string
            enum: [S, M, L, XL, XXL]
          minItems: 1
        colors:
          type: array
          items:
            type: string
          minItems: 1
        is_bestseller:
          type: boolean
          default: false
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          type: number
          minimum: 0
          maximum: 1000000
        description:
          type: string
          minLength: 10
          maxLength: 2000
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        sizes:
          type: array
          items:
            type: string
            enum: [S, M, L, XL, XXL]
          minItems: 1
        colors:
          type: array
          items:
            type: string
          minItems: 1
        is_bestseller:
          type: boolean
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ============================================
    # ORDER SCHEMAS
    # ============================================
...
    Error:
      type: object
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'